added many options, like autoload posts, navigation bar, single-page functions in the app
