+++ package.json
@@ -9,8 +9,9 @@
   "dependencies": {
     "axios": "^0.27.2",
     "core-js": "^3.8.3",
+    "travis": "^0.1.1",
     "vue": "^3.2.13",
-    "vue-router": "^4.0.3",
+    "vue-router": "^4.0.13",
     "vuex": "^4.0.0"
   },
   "devDependencies": {
