+++ src/components/PostItem.vue
@@ -1,11 +1,16 @@
 <template>
   <div class="post">
     <div>
-<!--      <div>{{ post.id }}</div>-->
+      <div>{{ post.id }}</div>
       <div><strong>Название:</strong> {{ post.title }}</div>
       <div><strong>Описание:</strong> {{ post.body }}</div>
     </div>
     <div class="post__btns">
+      <my-button
+          @click="$router.push(`/posts/${post.id}`)"
+      >
+        Открыть
+      </my-button>
       <my-button
            @click="$emit('remove', post)"
       >
@@ -34,5 +39,8 @@ export default {
   display: flex;
   align-self: center;
   justify-content: space-between;
+}
+.post__btns {
+  display: flex;
 }
 </style>
