+++ src/pages/About.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>Это страница абоут</h1>
+  </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style scoped>
+
+</style>
