+++ src/pages/Main.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    <h1>Это стартовая страница одностраничного приложения social-post</h1>
+    <p>Многие функции еще не готовы и не все реализованы.<br>
+    Сейчас работают все кнопки, но функционал есть только на кнопке
+      <router-link to="/posts">"Посты"</router-link>
+    </p>
+  </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style scoped>
+
+</style>
