+++ src/pages/PostIdPage.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>Это страница поста с ID = {{ $route.params.id }}</h1>
+  </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style scoped>
+
+</style>
