+++ src/pages/PostsPage.vue
@@ -1,0 +1,189 @@
+<template>
+  <div>
+    <h1>Страница с постами</h1>
+    <my-input
+        v-model="searchQuery"
+        placeholder="Поиск"
+    />
+    <div class="app__btns">
+      <my-button
+          @click="showDialog"
+      >
+        Создать пост
+      </my-button>
+      <my-select
+          v-model="selectedSort"
+          :options="sortOptions"
+      />
+    </div>
+    <my-dialog v-model:show="dialogVisible">
+      <post-form
+          @create="createPost"
+      />
+    </my-dialog>
+
+    <post-list
+        :posts="sortedAndSearchedPosts"
+        @remove="removePost"
+        v-if="!isPostLoading"
+    />
+    <div v-else>Идет загрузка...</div>
+    <div ref="observer" class="observer"></div>
+    <!--    <div class="page__wrapper">-->
+    <!--      <div-->
+    <!--          v-for="pageNumber in totalPages"-->
+    <!--          :key="pageNumber"-->
+    <!--          class="page"-->
+    <!--          :class="{-->
+    <!--          'current-page': page === pageNumber-->
+    <!--        }"-->
+    <!--          @click="changePage(pageNumber)"-->
+    <!--      >-->
+    <!--        {{ pageNumber }}-->
+    <!--      </div>-->
+    <!--    </div>-->
+  </div>
+
+</template>
+
+<script>
+import PostForm from "@/components/PostForm";
+import PostList from "@/components/PostList";
+import MyButton from "@/components/UI/MyButton";
+import axios from 'axios';
+import MySelect from "@/components/UI/MySelect";
+import MyDialog from "@/components/UI/MyDialog";
+import MyInput from "@/components/UI/MyInput";
+
+export default {
+  components: {
+    MyInput,
+    MyDialog,
+    MySelect,
+    MyButton,
+    PostList, PostForm
+  },
+  data() {
+    return {
+      posts: [],
+      dialogVisible: false,
+      isPostLoading: false,
+      selectedSort: '',
+      searchQuery: '',
+      page: 1,
+      limit: 10,
+      totalPages: 0,
+      sortOptions: [
+        {value: 'title', name: 'По названию'},
+        {value: 'body', name: 'По содержимому'}
+      ]
+    }
+  },
+  methods: {
+    createPost(post) {
+      this.posts.push(post);
+      this.dialogVisible = false;
+    },
+    removePost(post) {
+      this.posts = this.posts.filter(p => p.id !== post.id)
+    },
+    showDialog() {
+      this.dialogVisible = true;
+    },
+    async loadMorePosts() {
+      try {
+        this.page += 1;
+        const response = await axios.get('https://jsonplaceholder.typicode.com/posts', {
+          params: {
+            _page: this.page,
+            _limit: this.limit
+          }
+        });
+        this.totalPages = Math.ceil(response.headers['x-total-count'] / this.limit)
+        this.posts = [...this.posts, ...response.data]
+      } catch (e) {
+        alert('Ошибка')
+      }
+    },
+    async fetchPosts() {
+      try {
+
+        this.isPostLoading = true;
+        const response = await axios.get('https://jsonplaceholder.typicode.com/posts', {
+          params: {
+            _page: this.page,
+            _limit: this.limit
+          }
+        });
+        this.totalPages = Math.ceil(response.headers['x-total-count'] / this.limit)
+        this.posts = response.data; // перезаписываем текущие посты на те, что с сервера
+      } catch (e) {
+        alert('Ошибка')
+      } finally {
+        this.isPostLoading = false;
+      }
+    },
+    // changePage(pageNumber) { // Это для страниц 1,2,3...
+    //   this.page = pageNumber
+    // }
+  },
+  mounted() {
+    this.fetchPosts();
+
+    console.log(this.$refs.observer)
+    const options = {
+      rootMargin: '0px',
+      threshold: 1.0
+    }
+    const callback = (entries, observer) => {
+      /* Content excerpted, show below */
+      if (entries[0].isIntersecting && this.page < this.totalPages) {
+        this.loadMorePosts(this.$refs.observer)
+      }
+    };
+    const observer = new IntersectionObserver(callback, options);
+    observer.observe(this.$refs.observer);
+  },
+  computed: {
+    sortedPosts() {
+      return [...this.posts].sort((post1, post2) => post1[this.selectedSort]?.localeCompare(post2[this.selectedSort]))
+    },
+    sortedAndSearchedPosts() {
+      return this.sortedPosts.filter(post => post.title.toLowerCase().includes(this.searchQuery.toLowerCase()))
+    }
+  },
+  watch: {
+    // page() {
+    //   this.fetchPosts() // Подгружает посты с сервера
+    // }
+  }
+};
+</script>
+
+<style>
+
+.app__btns {
+  margin-top: 15px;
+  display: flex;
+  justify-content: space-between;
+}
+
+.page__wrapper {
+  display: flex;
+  margin-top: 15px;
+}
+
+.page {
+  border: 1px solid black;
+  padding: 10px;
+  cursor: pointer;
+}
+.current-page {
+  border: 2px solid teal;
+}
+
+.observer {
+  height: 30px;
+  background: green;
+}
+</style>
